Update cosmos deps and sync code changes
