+++ package.json
@@ -35,9 +35,7 @@
     "typescript": "^4.9.5"
   },
   "resolutions": {
-    "@cosmjs/cosmwasm-stargate": "^0.30.1",
-    "node-fetch": "2.6.7",
-    "@subql/types": "2.2.1-1"
+    "node-fetch": "2.6.7"
   },
   "scripts": {
     "build": "yarn workspaces foreach -ptA run build",
