+++ packages/node/package.json
@@ -20,9 +20,9 @@
   },
   "dependencies": {
     "@apollo/client": "3.5.8",
-    "@cosmjs/cosmwasm-stargate": "^0.30.1",
-    "@cosmjs/proto-signing": "^0.30.1",
-    "@cosmjs/stargate": "^0.30.1",
+    "@cosmjs/cosmwasm-stargate": "^0.32.3",
+    "@cosmjs/proto-signing": "^0.32.3",
+    "@cosmjs/stargate": "^0.32.3",
     "@nestjs/common": "^9.4.0",
     "@nestjs/core": "^9.4.0",
     "@nestjs/event-emitter": "^2.0.0",
@@ -32,7 +32,6 @@
     "@subql/common-cosmos": "workspace:*",
     "@subql/node-core": "^8.0.1",
     "@subql/types-cosmos": "workspace:*",
-    "cosmjs-types": "^0.7.0",
     "cron-converter": "^1.0.2",
     "eventemitter2": "^6.4.5",
     "lodash": "^4.17.21",
