+++ packages/node/src/indexer/api.service.ts
@@ -6,7 +6,7 @@ import { CosmWasmClient, IndexedTx } from '@cosmjs/cosmwasm-stargate';
 import { toHex } from '@cosmjs/encoding';
 import { Uint53 } from '@cosmjs/math';
 import { DecodeObject, GeneratedType, Registry } from '@cosmjs/proto-signing';
-import { Block, defaultRegistryTypes } from '@cosmjs/stargate';
+import { Block, defaultRegistryTypes, SearchTxQuery } from '@cosmjs/stargate';
 import {
   Tendermint37Client,
   toRfc3339WithNanoseconds,
@@ -23,6 +23,7 @@ import {
   getLogger,
   ConnectionPoolService,
   ApiService as BaseApiService,
+  IBlock,
 } from '@subql/node-core';
 import { CosmWasmSafeClient } from '@subql/types-cosmos/interfaces';
 import {
@@ -42,7 +43,7 @@ const logger = getLogger('api');
 
 @Injectable()
 export class ApiService
-  extends BaseApiService<CosmosClient, CosmosSafeClient, BlockContent[]>
+  extends BaseApiService<CosmosClient, CosmosSafeClient, IBlock<BlockContent>[]>
   implements OnApplicationShutdown
 {
   private fetchBlocksBatches = CosmosUtil.fetchBlocksBatches;
@@ -163,7 +164,7 @@ export class CosmosClient extends CosmWasmClient {
 
   // eslint-disable-next-line @typescript-eslint/require-await
   async txInfoByHeight(height: number): Promise<readonly IndexedTx[]> {
-    return this.searchTx({ height: height });
+    return this.searchTx(`tx.height=${height}`);
   }
 
   // eslint-disable-next-line @typescript-eslint/require-await
@@ -225,7 +226,7 @@ export class CosmosSafeClient
 
   // Deprecate
   async getBlock(): Promise<Block> {
-    const response = await this.forceGetTmClient().block(this.height);
+    const response = await this.forceGetCometClient().block(this.height);
     return {
       id: toHex(response.blockId.hash).toUpperCase(),
       header: {
@@ -243,21 +244,23 @@ export class CosmosSafeClient
 
   async validators(): Promise<readonly Validator[]> {
     return (
-      await this.forceGetTmClient().validators({
+      await this.forceGetCometClient().validators({
         height: this.height,
       })
     ).validators;
   }
 
-  async searchTx(): Promise<readonly IndexedTx[]> {
-    const txs: readonly IndexedTx[] = await this.safeTxsQuery(
+  async searchTx(query: SearchTxQuery): Promise<IndexedTx[]> {
+    const txs: IndexedTx[] = await this.safeTxsQuery(
       `tx.height=${this.height}`,
     );
     return txs;
   }
 
-  private async safeTxsQuery(query: string): Promise<readonly IndexedTx[]> {
-    const results = await this.forceGetTmClient().txSearchAll({ query: query });
+  private async safeTxsQuery(query: string): Promise<IndexedTx[]> {
+    const results = await this.forceGetCometClient().txSearchAll({
+      query: query,
+    });
     return results.txs.map((tx) => {
       return {
         txIndex: tx.index,
@@ -268,6 +271,7 @@ export class CosmosSafeClient
         tx: tx.tx,
         gasUsed: tx.result.gasUsed,
         gasWanted: tx.result.gasWanted,
+        msgResponses: [], // TODO can we get these?
         events: tx.result.events.map((evt) => ({
           ...evt,
           attributes: evt.attributes.map((attr) => ({
