+++ packages/node/src/indexer/blockDispatcher/cosmos-block-dispatcher.ts
@@ -2,7 +2,8 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import { IBlockDispatcher } from '@subql/node-core';
+import { BlockContent } from '../types';
 
-export interface ICosmosBlockDispatcher extends IBlockDispatcher {
+export interface ICosmosBlockDispatcher extends IBlockDispatcher<BlockContent> {
   init(onDynamicDsCreated: (height: number) => Promise<void>): Promise<void>;
 }
