+++ packages/node/src/indexer/dictionary/v1/index.ts
@@ -1,0 +1,4 @@
+// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// SPDX-License-Identifier: GPL-3.0
+
+export * from './dictionaryV1';
