+++ packages/node/src/indexer/fetch.module.ts
@@ -24,7 +24,7 @@ import {
   WorkerBlockDispatcherService,
 } from './blockDispatcher';
 import { CosmosClientConnection } from './cosmosClient.connection';
-import { DictionaryService } from './dictionary.service';
+import { DictionaryService } from './dictionary/dictionary.service';
 import { DsProcessorService } from './ds-processor.service';
 import { DynamicDsService } from './dynamic-ds.service';
 import { FetchService } from './fetch.service';
@@ -67,7 +67,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
         unfinalizedBlocks: UnfinalizedBlocksService,
         connectionPoolState: ConnectionPoolStateManager<CosmosClientConnection>,
       ) =>
-        nodeConfig.workers !== undefined
+        nodeConfig.workers
           ? new WorkerBlockDispatcherService(
               nodeConfig,
               eventEmitter,
@@ -119,22 +119,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
     ConnectionPoolService,
     IndexingBenchmarkService,
     PoiBenchmarkService,
-    {
-      provide: DictionaryService,
-      useFactory: async (
-        project: SubqueryProject,
-        nodeConfig: NodeConfig,
-        eventEmitter: EventEmitter2,
-      ) => {
-        const dictionaryService = await DictionaryService.create(
-          project,
-          nodeConfig,
-          eventEmitter,
-        );
-        return dictionaryService;
-      },
-      inject: ['ISubqueryProject', NodeConfig, EventEmitter2],
-    },
+    DictionaryService,
     SandboxService,
     DsProcessorService,
     DynamicDsService,
