+++ packages/node/src/indexer/unfinalizedBlocks.service.ts
@@ -9,20 +9,9 @@ import {
   StoreCacheService,
   mainThreadOnly,
 } from '@subql/node-core';
+import { cosmosBlockToHeader } from '../utils/cosmos';
 import { ApiService } from './api.service';
 import { BlockContent } from './types';
-
-/*
- * Cosmos has instant finalization, there is also no rpc method to get a block by hash
- * To get around this we use blockHeights as hashes
- */
-export function cosmosBlockToHeader(blockHeight: number): Header {
-  return {
-    blockHeight: blockHeight,
-    blockHash: blockHeight.toString(),
-    parentHash: (blockHeight - 1).toString(),
-  };
-}
 
 @Injectable()
 export class UnfinalizedBlocksService extends BaseUnfinalizedBlocksService<BlockContent> {
@@ -32,11 +21,6 @@ export class UnfinalizedBlocksService extends BaseUnfinalizedBlocksService<Block
     storeCache: StoreCacheService,
   ) {
     super(nodeConfig, storeCache);
-  }
-
-  @mainThreadOnly()
-  protected blockToHeader(block: BlockContent): Header {
-    return cosmosBlockToHeader(block.block.header.height);
   }
 
   @mainThreadOnly()
