+++ packages/node/src/utils/cosmos.ts
@@ -13,7 +13,7 @@ import {
   TxData,
   Event,
 } from '@cosmjs/tendermint-rpc/build/tendermint37';
-import { getLogger } from '@subql/node-core';
+import { IBlock, getLogger, Header } from '@subql/node-core';
 import {
   CosmosEventFilter,
   CosmosMessageFilter,
@@ -25,7 +25,6 @@ import {
   CosmosTxFilter,
 } from '@subql/types-cosmos';
 import { isObjectLike } from 'lodash';
-import { isLong } from 'long';
 import { CosmosClient } from '../indexer/api.service';
 import { BlockContent } from '../indexer/types';
 
@@ -71,17 +70,9 @@ export function filterMessageData(
   }
   if (filter.values) {
     for (const key in filter.values) {
-      let decodedMsgData = key
+      const decodedMsgData = key
         .split('.')
         .reduce((acc, curr) => acc[curr], data.msg.decodedMsg);
-
-      //stringify Long for equality check
-      if (isLong(decodedMsgData)) {
-        decodedMsgData =
-          typeof filter.values[key] === 'number'
-            ? decodedMsgData.toNumber()
-            : decodedMsgData.toString();
-      }
 
       if (filter.values[key] !== decodedMsgData) {
         return false;
@@ -319,10 +310,29 @@ export function wrapEvent(
   return events;
 }
 
+/*
+ * Cosmos has instant finalization, there is also no rpc method to get a block by hash
+ * To get around this we use blockHeights as hashes
+ */
+export function cosmosBlockToHeader(blockHeight: number): Header {
+  return {
+    blockHeight: blockHeight,
+    blockHash: blockHeight.toString(),
+    parentHash: (blockHeight - 1).toString(),
+  };
+}
+
+export function formatBlockUtil<B extends BlockContent>(block: B): IBlock<B> {
+  return {
+    block,
+    getHeader: () => cosmosBlockToHeader(block.block.header.height),
+  };
+}
+
 export async function fetchBlocksBatches(
   api: CosmosClient,
   blockArray: number[],
-): Promise<BlockContent[]> {
+): Promise<IBlock<BlockContent>[]> {
   const blocks = await fetchCosmosBlocksArray(api, blockArray);
   return blocks.map(([blockInfo, blockResults]) => {
     try {
@@ -331,7 +341,9 @@ export async function fetchBlocksBatches(
         `txInfos doesn't match up with block (${blockInfo.block.header.height}) transactions expected ${blockInfo.block.txs.length}, received: ${blockResults.results.length}`,
       );
 
-      return new LazyBlockContent(blockInfo, blockResults, api);
+      return formatBlockUtil(
+        new LazyBlockContent(blockInfo, blockResults, api),
+      );
     } catch (e) {
       logger.error(
         e,
