+++ packages/types/package.json
@@ -16,9 +16,9 @@
     "/dist"
   ],
   "dependencies": {
-    "@cosmjs/cosmwasm-stargate": "^0.30.1",
-    "@cosmjs/proto-signing": "^0.30.1",
-    "@cosmjs/stargate": "^0.30.1",
+    "@cosmjs/cosmwasm-stargate": "^0.32.3",
+    "@cosmjs/proto-signing": "^0.32.3",
+    "@cosmjs/stargate": "^0.32.3",
     "@subql/types-core": "^0.6.0"
   }
 }
